Add workaround for failing yum conveyor test
There's code in singularity that assumes that if rpm is installed on the
system it will behave like rpm on a rpm-based distribution. Debian and
derivatives configure rpm in a way that makes sense in those systems,
but it's not the same as in rpm-based distributions.

The tests for the yum conveyor actually check that rpm behaves in that
particular way, and FAIL otherwise.

Work around this issue by adding macros to satisfy this assumption.

See https://github.com/sylabs/singularity/issues/241, too.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>